EDM-286 Showing default device labels in Approve Device Modal

+++ libs/ui-components/src/components/form/LabelsField.tsx
@@ -1,30 +1,30 @@
 import * as React from 'react';
 import { useField } from 'formik';
-import { Button, Label, LabelGroup } from '@patternfly/react-core';
-import { TimesIcon } from '@patternfly/react-icons/dist/js/icons/times-icon';
+import { Label, LabelGroup } from '@patternfly/react-core';
 
 import { FlightCtlLabel } from '../../types/extraTypes';
 import EditableLabelControl from '../common/EditableLabelControl';
-import { useTranslation } from '../../hooks/useTranslation';
 import ErrorHelperText, { DefaultHelperText } from './FieldHelperText';
 
 type LabelsFieldProps = {
   name: string;
-  isEditable?: boolean;
+  isFormEditable?: boolean;
   addButtonText?: string;
   helperText?: React.ReactNode;
   onChangeCallback?: (newLabels: FlightCtlLabel[], hasErrors: boolean) => void;
 };
+
+const maxWidthDefaultLabel = '18ch'; // Can fit more chars as it doesn't have a "Close" button
+const maxWidthNonDefaultLabel = '16ch'; // Can fit less chars due to the "Close" button
 
 const LabelsField: React.FC<LabelsFieldProps> = ({
   name,
   onChangeCallback,
   addButtonText,
   helperText,
-  isEditable = true,
+  isFormEditable = true,
 }) => {
   const [{ value: labels }, meta, { setValue: setLabels }] = useField<FlightCtlLabel[]>(name);
-  const { t } = useTranslation();
   const updateLabels = async (newLabels: FlightCtlLabel[]) => {
     const errors = await setLabels(newLabels, true);
     const hasErrors = Object.keys(errors || {}).length > 0;
@@ -33,7 +33,7 @@ const LabelsField: React.FC<LabelsFieldProps> = ({
   };
 
   const onDelete = async (_ev: React.MouseEvent<Element, MouseEvent>, index: number) => {
-    if (!isEditable) {
+    if (!isFormEditable) {
       return;
     }
     const newLabels = [...labels];
@@ -67,33 +67,45 @@ const LabelsField: React.FC<LabelsFieldProps> = ({
     <>
       <LabelGroup
         numLabels={5}
-        isEditable={isEditable}
+        isEditable={isFormEditable}
         addLabelControl={
           <EditableLabelControl
             defaultLabel="key=value"
             addButtonText={addButtonText}
             onAddLabel={onAdd}
-            isEditable={isEditable}
+            isEditable={isFormEditable}
           />
         }
       >
-        {labels.map(({ key, value }, index) => (
-          <Label
-            key={index}
-            id={`${index}`}
-            closeBtn={
-              isEditable ? undefined : (
-                <Button variant="plain" aria-label={t('Delete')} isDisabled icon={<TimesIcon />} />
-              )
-            }
-            onClose={(e) => onDelete(e, index)}
-            onEditCancel={(_, prevText) => onEdit(index, prevText)}
-            onEditComplete={(_, newText) => onEdit(index, newText)}
-            isEditable={isEditable}
-          >
-            {value ? `${key}=${value}` : key}
-          </Label>
-        ))}
+        {labels.map(({ key, value, isDefault }, index) => {
+          const text = value ? `${key}=${value}` : key;
+          const elKey = `${key}__${index}`;
+          if (isDefault) {
+            return (
+              <Label key={elKey} textMaxWidth={maxWidthDefaultLabel}>
+                {text}
+              </Label>
+            );
+          }
+
+          const closeButtonProps = !isFormEditable && { isDisabled: true };
+          const isLabelEditable = isFormEditable && !isDefault;
+          return (
+            <Label
+              key={elKey}
+              textMaxWidth={maxWidthNonDefaultLabel}
+              closeBtnProps={closeButtonProps}
+              onClose={(e) => onDelete(e, index)}
+              onEditCancel={(_, prevText) => onEdit(index, prevText)}
+              onEditComplete={(_, newText) => onEdit(index, newText)}
+              /* Add a basic tooltip as the PF tooltip doesn't work for editable labels */
+              title={isLabelEditable ? text : undefined}
+              isEditable={isLabelEditable}
+            >
+              {text}
+            </Label>
+          );
+        })}
       </LabelGroup>
       <DefaultHelperText helperText={helperText} />
       <ErrorHelperText meta={meta} touchRequired={false} />
